+++ www/front_src/src/Resources/Graph/Performance/index.tsx
@@ -4,11 +4,9 @@ import {
   useEffect,
   useRef,
   useState,
-  memo,
 } from 'react';
 
-import { AddSVGProps } from '@visx/shape/lib/types';
-import { Responsive, Shape } from '@visx/visx';
+import { Responsive } from '@visx/visx';
 import { useAtomValue } from 'jotai/utils';
 import {
   add,
@@ -30,7 +28,7 @@ import {
 } from 'ramda';
 import { useTranslation } from 'react-i18next';
 
-import { alpha, Skeleton, Theme, Typography, useTheme } from '@mui/material';
+import { Skeleton, Theme, Typography } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 
 import {
@@ -40,16 +38,16 @@ import {
   useRequest,
 } from '@centreon/ui';
 
-import { labelNoDataForThisPeriod } from '../../translatedLabels';
-import { TimelineEvent } from '../../Details/tabs/Timeline/models';
-import { Resource, ResourceType } from '../../models';
 import { CommentParameters } from '../../Actions/api';
+import { selectedResourcesDetailsAtom } from '../../Details/detailsAtoms';
 import { ResourceDetails } from '../../Details/models';
 import {
   CustomTimePeriod,
   CustomTimePeriodProperty,
 } from '../../Details/tabs/Graph/models';
-import { selectedResourcesDetailsAtom } from '../../Details/detailsAtoms';
+import { TimelineEvent } from '../../Details/tabs/Timeline/models';
+import { Resource, ResourceType } from '../../models';
+import { labelNoDataForThisPeriod } from '../../translatedLabels';
 
 import { CustomFactorsData } from './AnomalyDetection/models';
 import Graph from './Graph';
@@ -146,23 +144,6 @@ const useStyles = makeStyles<Theme, MakeStylesProps>((theme) => ({
 
 const shiftRatio = 2;
 
-interface BarProps {
-  className?: string;
-  innerRef?: React.Ref<SVGRectElement>;
-  open: boolean;
-}
-
-const Bar = ({
-  open,
-  ...restProps
-}: AddSVGProps<BarProps, SVGRectElement>): JSX.Element | null => {
-  if (!open) {
-    return null;
-  }
-
-  return <Shape.Bar {...restProps} />;
-};
-
 const PerformanceGraph = ({
   endpoint,
   graphHeight,
@@ -191,7 +172,6 @@ const PerformanceGraph = ({
     graphHeight,
   });
   const { t } = useTranslation();
-  const theme = useTheme();
 
   const [timeSeries, setTimeSeries] = useState<Array<TimeValue>>([]);
   const [lineData, setLineData] = useState<Array<LineModel>>();
@@ -424,11 +404,6 @@ const PerformanceGraph = ({
 
   const displayTimeValues = not(isListingGraphOpen) || isDisplayedInListing;
 
-  const propsAreEqual = (prevProps, nextProps): boolean =>
-    equals(prevProps, nextProps);
-
-  const MemoizedBar = memo(Bar, propsAreEqual);
-
   return (
     <div
       className={classes.container}
@@ -467,6 +442,7 @@ const PerformanceGraph = ({
               displayEventAnnotations={displayEventAnnotations}
               displayTimeValues={displayTimeValues}
               height={height}
+              interactWithGraph={interactWithGraph}
               lines={displayedLines}
               loading={
                 not(resourceDetailsUpdated) && sendingGetGraphDataRequest
@@ -479,48 +455,7 @@ const PerformanceGraph = ({
               width={width}
               xAxisTickFormat={xAxisTickFormat}
               onAddComment={onAddComment}
-            >
-              {({
-                closeTooltip,
-                displayAddCommentTooltip,
-                displayTooltip,
-                displayZoomPreview,
-                BarHeight,
-                graphWidth,
-                style,
-                zoomBarWidth,
-                zoomBoundaries,
-              }): JSX.Element => (
-                <>
-                  {zoomBarWidth && (
-                    <MemoizedBar
-                      open
-                      fill={alpha(theme.palette.primary.main, 0.2)}
-                      height={graphHeight}
-                      stroke={alpha(theme.palette.primary.main, 0.5)}
-                      width={zoomBarWidth}
-                      x={zoomBoundaries?.start || 0}
-                      y={0}
-                    />
-                  )}
-                  {displayAddCommentTooltip && (
-                    <MemoizedBar
-                      className={style?.overlay}
-                      fill="transparent"
-                      height={BarHeight}
-                      open={interactWithGraph}
-                      width={graphWidth}
-                      x={0}
-                      y={0}
-                      onMouseDown={displayZoomPreview}
-                      onMouseLeave={closeTooltip}
-                      onMouseMove={displayTooltip}
-                      onMouseUp={displayAddCommentTooltip}
-                    />
-                  )}
-                </>
-              )}
-            </Graph>
+            />
           )}
         </Responsive.ParentSize>
       </div>
